Added implementation of environment.
RL agent converges in the simplified problem of controlling the temperature.
Next step is to add the efficiency requirement.
Added script to graph the series.

+++ aer/utils.py
@@ -26,6 +26,6 @@ def read_series(source_path: str) -> Array[float, -1]:
         stamp = parse_datetime(raw_series[i][0])
         raw_series[i][0] = datetime.datetime.strptime(stamp, "%Y-%m-%d %H:%M:%S")
 
-    series: Array[-1] = raw_series[:, 1]
+    series: Array[-1] = raw_series[:, 1].astype(float)
 
     return series
